Fix compiler warnings (float/double conversion)

+++ OBSlider/OBSlider.m
@@ -109,7 +109,7 @@
         CGFloat trackingOffset = currentLocation.x - previousLocation.x;
         
         // Find the scrubbing speed that curresponds to the touch's vertical offset
-        CGFloat verticalOffset = fabsf(currentLocation.y - self.beganTrackingLocation.y);
+        CGFloat verticalOffset = fabs(currentLocation.y - self.beganTrackingLocation.y);
         NSUInteger scrubbingSpeedChangePosIndex = [self indexOfLowerScrubbingSpeed:self.scrubbingSpeedChangePositions forOffset:verticalOffset];        
         if (scrubbingSpeedChangePosIndex == NSNotFound) {
             scrubbingSpeedChangePosIndex = [self.scrubbingSpeeds count];
@@ -117,15 +117,15 @@
         self.scrubbingSpeed = [[self.scrubbingSpeeds objectAtIndex:scrubbingSpeedChangePosIndex - 1] floatValue];
          
         CGRect trackRect = [self trackRectForBounds:self.bounds];
-        self.realPositionValue = self.realPositionValue + (self.maximumValue - self.minimumValue) * (trackingOffset / trackRect.size.width);
+        self.realPositionValue = self.realPositionValue + (self.maximumValue - self.minimumValue) * (float)(trackingOffset / trackRect.size.width);
 		
-		CGFloat valueAdjustment = self.scrubbingSpeed * (self.maximumValue - self.minimumValue) * (trackingOffset / trackRect.size.width);
-		CGFloat thumbAdjustment = 0.0f;
+		float valueAdjustment = self.scrubbingSpeed * (self.maximumValue - self.minimumValue) * (float)(trackingOffset / trackRect.size.width);
+		float thumbAdjustment = 0.0f;
         if ( ((self.beganTrackingLocation.y < currentLocation.y) && (currentLocation.y < previousLocation.y)) ||
              ((self.beganTrackingLocation.y > currentLocation.y) && (currentLocation.y > previousLocation.y)) )
             {
             // We are getting closer to the slider, go closer to the real location
-			thumbAdjustment = (self.realPositionValue - self.value) / (1 + fabsf(currentLocation.y - self.beganTrackingLocation.y));
+			thumbAdjustment = (self.realPositionValue - self.value) / (float)(1 + fabs(currentLocation.y - self.beganTrackingLocation.y));
         }
 		self.value += valueAdjustment + thumbAdjustment;
